Modified all code writers to permit their output to be written to a specific directory instead of the directory containing the .def file. This allows for easier management of generated files. All the writers take a -d option followed by the name of the output directory. This should be the root of a tree parallelling the source tree.
Also, the .gen files are now named after the corresponding .def file
instead of having a fixed name. This cures the problem that
directories having multiple .def files only list the files generated
from the last .def file that was processed. It also permits the .gen
file to be used for an "up-to-date" check against the .def file.

All writers extend a base class that supports the target directory and
.gen file name. Probably other functions could be moved there as well,
but this hasn't been done.