Formatting fixes: 2 or more hyphens seems to confuse the parser.
Also, fix up the check-sources target to correctly set the
jarfile.uptodate property to true if module.jar is dated same or later
than every .java file in the src directory.

Note that this uses module.jar, not module-<date>.jar
Also note that this property, although now correctly set, is not in
fact used anywhere.