Fixed bug 2502, "Create unified source view".
The "UnifySourceScripter.java" generates a script that creates
a symbolically-linked view of the Cougaar sources.  This is more
like documentation than release code, since it's an optional step
after the source has been compiled (like javadocs) and does not
affect runtime.

I've also checked in the generated "unify_script.sh".  It should be
regenerated whenever the modules change (ideally as part of the build).